requirements.txt
python-dotenv==0.19.1
aiogram~=2.25.1
openai~=0.27.8